+++ src/main/java/dev/shadowsoffire/placebo/menu/MenuUtil.java
@@ -53,6 +53,9 @@ public class MenuUtil {
         return InteractionResult.CONSUME;
     }
 
+
+
+
     /**
      * IIntArray can only send shorts, so we need to split int values in two.
      *